@@ -13,7 +13,6 @@
     "start": "./node_modules/.bin/webpack serve --config ./config/webpack.dev.js --mode development"
   },
   "devDependencies": {
-    "@types/styled-components": "^5.1.3",
     "@typescript-eslint/eslint-plugin": "^5.6.0",
     "@typescript-eslint/parser": "^5.6.0",
     "clean-webpack-plugin": "^4.0.0",
@@ -38,13 +37,14 @@
     "webpackbar": "^5.0.0-3"
   },
   "dependencies": {
-    "@types/react": "^17.0.0",
-    "@types/react-dom": "^17.0.0",
+    "@types/react": "^18.0.5",
+    "@types/react-dom": "^18.0.1",
+    "@types/styled-components": "^5.1.25",
     "@types/uuid": "^3.4.9",
     "normalize.css": "^8.0.1",
-    "react": "^17.0.1",
-    "react-dom": "^17.0.1",
-    "styled-components": "^5.2.0",
+    "react": "^18.0.0",
+    "react-dom": "^18.0.0",
+    "styled-components": "^5.3.5",
     "styled-normalize": "^8.0.7",
     "typewriter-effect": "^2.18.2",
     "uuid": "^3.4.0"
